Did some modofications on the LineDirectedGraphs class
