+++ jgrapht-core/src/test/java/org/jgrapht/LineDirectedGraphsTest.java
@@ -1,4 +1,21 @@
-package org.jgrapht.own;
+/*
+ * (C) Copyright 2018-2019, by Arica Chakraborty and Contributors.
+ *
+ * JGraphT : a free Java graph-theory library
+ *
+ * This program and the accompanying materials are dual-licensed under
+ * either
+ *
+ * (a) the terms of the GNU Lesser General Public License version 2.1
+ * as published by the Free Software Foundation, or (at your option) any
+ * later version.
+ *
+ * or (per the licensee's choosing)
+ *
+ * (b) the terms of the Eclipse Public License v1.0 as published by
+ * the Eclipse Foundation.
+ */
+package org.jgrapht;
 
 import org.jgrapht.Graph;
 import org.jgrapht.graph.DefaultEdge;
